inventory: Don't add *_remove_from_domain due to delege_to issue
If "ipaserver_remove_from_domain" or "ipareplica_remove_from_domain" are
used during server undeployment, the playbook fails whil trying to
connect to the server through SSH instead of the podman connector.

It still not clean why issue with happens in "delegate_to" tasks, and
until it is sorted out, the option will not be added automatically to
the generated inventory.