Add support for custom containerfiles
A new commaand line argument ('-f') and a new configuration variable
are added so that custom containerfiles can be used in the creation of
the containers.

The '-f CONTAINERFILE' must be used so that the custom containerfile
is included in the data files for the generated configuration.

To use the custom containerfile in the creation of the node containers,
the 'distro' variable can be used for the whole deployment, setting the
default 'distro' for the deployment, or may be overriden on each node
definition.