Add suport for podman '--dns' throught the 'dns' option
When using a custom container file there might be issues with DNS due to
the default '/etc/resolv.conf' file generated for the containers, this
is why the provided container files add some helper tools to fix this
configuration and allow replicas and clients to be deployed.

To circunvent this issue, podman's argument `--dns` might be used,
relieving some of the original problems, and allowing client nodes to be
deployed (altought it does not work for replica deployment).

The option 'dns' was added to the configuration script, and can be used
as a deployment or a node variable. The value will be ignored on the
first server of the deployment, as using 'dns' may cause external hosts
to be unacessible, and it is supposed that the first server need
external access to download the software packages. With 'dns' and if the
first server has DNS role installed, the clients can use this server as
nameserver and can resolve external names, making installation of
software packages possible.

A new example with custom containers, muiltple deployments, and 'dns'
setting is available:

    examples/custom-containerfile-with-dns.yml