inventory: Fix variables for multiple deployments
When setting variables in groups that are repeated in different
deployments, Ansible overwrite the last variables with the values form
the last defined deployment. As this is the case for any deployment, as
ipaserver, ipareplicas and ipaclients groups are in all deployments,
this causes all deployment to share the same variables, unless they are
overriden explicitly for the node.

The fix was to remove the group variables and define the variables for
each node individually. This results in a bigger and harder to read
file, but allows the deployments to hold different values for variables
with the same name.