Convert container names to FQDN
The container names are defined as free strings where the only rules
that apply (but not enforced) are Podman rules for container names.

Usually the names for the containers contain a single word, as they
represent services in a compose, but in the case of a FreeIPA cluster,
it makes sense to have the container names to mimic the hostnames,
specially in multi-deployment environments.

This change add support for the global configuration 'container_fqdn'.
When set to 'true', if  the container is a single word, the name is
appended by the deployment domain, with the container name becoming
"<name>.<domain>" in the same way that the hostnames are derived.

By default, 'container_fqdn' is 'false'.

With this change, when using 'container_fqdn: true' the podman network
nameserver resolves the names correctly and replicas can be deployed
even with multiple-deployment scenarios.