Allow adding Ansible playbooks to output 'playbooks' directory
This patch adds CLI parameter '-p/--playbook' to create a list of files
to be added to the output 'playbooks' directory.

Both files and directories can be used as arguments, and directories are
recursively searched for '.yml' and '.yaml' files.

The output directory is flat, so files with the same file name will be
overwriten.