Add support for 'subnet' definition
One of the issues of using a random network for assigning IP address is
that one need to manually remove the network configuration generated by
podman and rebuild the containers. Also, the number of IP addresses were
limited to 250.

By allowing the use of a CIDR a user can define the base address for the
hosts and the number of available IP addresses. The default value used
for 'subnet' is "192.168.159.0".

A new example is provided:

    examples/custom-ip-address.yml