network: Simplify network configuration code
Since we are explicitly setting "name" for networks, we can always use
the same name as the network refernce in the compose file (and in the
internal structure), simplifying the way we handle network internally.

The resulting runtime compose will have the same network configuration
as in previous versions, so nothing should change for the user.