Add support for external networks.
The docker compose spec does not allow disabling the internal DNS
nameserver for user defined networks, or setting the network DNS
resolver. This may cause conflicts when complete control of DNS
zones (direct and reverse) is needed.

By allowing the usage of a pre-existing network, the network can be
created with the required configuration, allowing networks with specific
nameservers, or with the internal DNS resolver disabled.

To use an external network, first create the network with:

    podman network create --subnet "192.168.12.0/24" my_external_network

When using external networks, the 'subnet' is required.

A new feature description is available:

    features/external_network.feature

Signed-off-by: Rafael Guterres Jeffman <rjeffman@redhat.com>