feature: Provided patch to /etc/hosts file
Context:
  In order to be able to access the cluster network from the host,
  As a developer
  I want to have a patch file to apply to /etc/hosts

Often there's need to access the cluster nodes through their IP
addresses and it requires obtaining the node IP address and adding the
proper entry to /etc/hosts files on the host machine.

This features provides a 'hosts' file in the output directory with all
the entries for the hosts in the created cluster.