Add support for nodes external to the IPA deployment
Many useful testing environments for IPA include hosts that are not part
of the IPA deployment. External nameservers, AD DC servers, external
identity providers, and others should be configured in the compose, and
be available in the inventory file, but only as external hosts.

A new top level directory was created, "external", that allows the
configuration of external hosts.

An example of a DNS nameserver is provided, and as its use is very close
to the IPA deployment, a basic configuration is provided through
ipalab-config, making in easier to have the proper configuration
running (DNS is hard enough on its own).

Signed-off-by: Rafael Guterres Jeffman <rjeffman@redhat.com>