Implement external nodes as plugins
Extending ipalab-config with external nodes required changes to the
configuration (compose and inventory) code. By treating the external
nodes as plugins, with the a well defined list of requirements, adding
new roles for external nodes is easier and more self contained.

The requirements to add a new external node role is to add a module to
'ipalab_config.external_role' with the name of the role, e.g.
'ipalab_config.external_role.addc', with a dict variable named
'base_config' containing the default configuration for the role, and an
implementation for the function:

    def gen_config(lab_config, base_dir, node, options):

Where, 'lab_config' is whole configuration derived from the given
configuration file, 'base_dir' is the base directory for the
configuration output, 'node' is the external node configuration, and
'options' is the user defined options (can be set to None, no option is
set).

Necessary files for the role can be added to 'ipalab_config/data'.

Signed-off-by: Rafael Guterres Jeffman <rjeffman@redhat.com>