inventory: Use rootless containers safe idrange
When using rootless containers, there's a limit on subuid/subgid
available for the container, and not limiting the default idrange on the
IPA deployment causes several issues and raises some limitations on what
can be done with the containerized environment.

This change modifies the default behavior to limit the default idarange
by setting idstart/idmax and the rid bases as variables for the server
deployment when using ansible-freeipa. It has no effect if the server is
deployed using the CLI installer.

A new per-node option 'no_limit_uid' is availabel and is only evaluated
for the first server of a deployment, it set to 'true', the previous
behavior, with no idrange configuration, will be used.

Signed-off-by: Rafael Guterres Jeffman <rjeffman@redhat.com>