compose: Enhanced network configuration
When using a network that requires some more elaborate configuration,
ipalab-config relied on the network to be previously created.

By allowing some configuration on the network, the usage of the
environment is simpler as podman-compose can create the network with the
given parameters.

As this uses 'x-podman.dns_disable' and 'x-podman.dns' extensions, it
requires that podman-compose version is above 1.3.0, and, as of late
April 2025, such a release does not exist, and if one want to use this
feature podman-compose must be installed from the official repository:

  pip install git+https://github.com/containers/podman-compose@main

Signed-off-by: Rafael Guterres Jeffman <rjeffman@redhat.com>