compose: Force node hosntname and IP in extra_hosts
When deploying an FreeIPA server, the procedure may fail when
configuring the CA if name is not resolvable. By ensuring that the host
IP address can be resolved through /etc/hosts allow deployment to
succeed.

The symptom of the issue as a failure in the deployment with no log
entry to debug, where 389ds crashes and dogtag fails to spawn.

Signed-off-by: Rafael Guterres Jeffman <rjeffman@redhat.com>